backend/requirements.txt
# Core
Django==5.2
djangorestframework==3.16.0
django-cors-headers==4.7.0
django-filter==25.1
psycopg2-binary==2.9.10
gunicorn==21.2.0
python-dotenv==1.0.1

# Soulseek (slskd API client)
requests==2.32.3

# Matching engine
rapidfuzz==3.11.0

# Track identification
pyacoustid==1.3.0

# Audio metadata
mutagen==1.47.0

# Scraping
beautifulsoup4==4.12.3
lxml==5.3.0
